+++ gen/com/yanncebron/m68kplugin/lang/psi/directive/M68kMacroCallParameter.java
@@ -27,12 +27,15 @@ import com.yanncebron.m68kplugin.lang.psi.M68kAdmApi;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmArd;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmAri;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmCcr;
+import com.yanncebron.m68kplugin.lang.psi.M68kAdmDfc;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmDrd;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmImm;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmPcd;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmPci;
+import com.yanncebron.m68kplugin.lang.psi.M68kAdmSfc;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmSr;
 import com.yanncebron.m68kplugin.lang.psi.M68kAdmUsp;
+import com.yanncebron.m68kplugin.lang.psi.M68kAdmVbr;
 
 public interface M68kMacroCallParameter extends M68kPsiElement {
 
@@ -61,6 +64,9 @@ public interface M68kMacroCallParameter extends M68kPsiElement {
   M68kAdmCcr getAdmCcr();
 
   @Nullable
+  M68kAdmDfc getAdmDfc();
+
+  @Nullable
   M68kAdmDrd getAdmDrd();
 
   @Nullable
@@ -73,9 +79,15 @@ public interface M68kMacroCallParameter extends M68kPsiElement {
   M68kAdmPci getAdmPci();
 
   @Nullable
+  M68kAdmSfc getAdmSfc();
+
+  @Nullable
   M68kAdmSr getAdmSr();
 
   @Nullable
   M68kAdmUsp getAdmUsp();
+
+  @Nullable
+  M68kAdmVbr getAdmVbr();
 
 }